LayoutTests/media/modern-media-controls/tracks-support/sorted-by-user-preferred-languages-expected.txt

This test checks that tracks are sorted based on the user's preferred languages.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS shadowRoot.querySelector('button.tracks') became different from null
PASS shadowRoot.querySelector('button.tracks').getBoundingClientRect().width became different from 0
Tapping tracks button...
Subtitles: [
  "Off",
  "Auto (Recommended)",
  "label-en Captions",
  "Descriptions",
  "label-en",
  "Spanish Captions",
  "Descriptions (Spanish)",
  "Spanish",
  "Chinese (China mainland)",
  "label-fr (French) Captions",
  "Descriptions (French)",
  "label-fr (French)",
  "German Captions",
  "Descriptions (German)",
  "German"
]
PASS successfullyParsed is true

TEST COMPLETE
